Task 2 : The Rule Of Five
In file chatbot.h / chatbot.cpp, make changes to the class ChatBot
such that it complies with the Rule of Five.
Make sure to properly allocate / deallocate memory resources on the
heap and also copy member data where it makes sense to you.
In each of the methods (e.g. the copy constructor),
print a string of the type „ChatBot Copy Constructor“ to the
console so that you can see which method is called in later examples.